requirements.txt
streamlit>=1.35.0
requests>=2.31.0
pandas>=2.0.0
google-api-python-client>=2.130.0
google-auth>=2.29.0
